requirements.txt
async-timeout==4.0.2
boto3==1.26.96
botocore==1.29.96
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
redis==4.5.4
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.19
